+++ Saucier720-app/src/app/app.component.html
@@ -1,8 +1,7 @@
-
-    <div class="topnav">
-        <a routerLink="Pantry">Pantry</a>
-        <a routerLink="List">List</a>
-        <a routerLink="Deals">Deals</a>
-        <a routerLink="Recipes">Recipes</a>
-      </div>
+<div class="topnav"><app class="title"></app>
+  <a routerLink="Pantry">Pantry</a>
+  <a routerLink="List">List</a>
+  <a routerLink="Deals">Deals</a>
+  <a routerLink="Recipes">Recipes</a>
+</div>
 <router-outlet></router-outlet>
