+++ Saucier720-app/src/app/testing/app.component.spec.ts
@@ -1,5 +1,5 @@
 import { TestBed } from '@angular/core/testing';
-import { AppComponent } from './app.component';
+import { AppComponent } from '../app.component';
 
 describe('AppComponent', () => {
   beforeEach(async () => {
