+++ Saucier720-app/src/app/testing/deals.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { DealsComponent } from './deals.component';
+import { DealsComponent } from '../deals/deals.component';
 
 describe('DealsComponent', () => {
   let component: DealsComponent;
